# lib/api-zod/src/generated/types/article.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArticlePillar } from "./articlePillar";

export interface Article {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  pillar: ArticlePillar;
  /** @nullable */
  coverImage: string | null;
  published: boolean;
  /** @nullable */
  readingTime: number | null;
  createdAt: Date;
  updatedAt: Date;
}