# lib/api-zod/src/generated/types/createArticleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateArticleBodyPillar } from "./createArticleBodyPillar";

export interface CreateArticleBody {
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  pillar: CreateArticleBodyPillar;
  /** @nullable */
  coverImage?: string | null;
  published?: boolean;
  /** @nullable */
  readingTime?: number | null;
}