# lib/api-zod/src/generated/types/updateArticleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateArticleBodyPillar } from "./updateArticleBodyPillar";

export interface UpdateArticleBody {
  title?: string;
  slug?: string;
  excerpt?: string;
  content?: string;
  pillar?: UpdateArticleBodyPillar;
  /** @nullable */
  coverImage?: string | null;
  published?: boolean;
  /** @nullable */
  readingTime?: number | null;
}