# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyCategory } from "./updateProjectBodyCategory";

export interface UpdateProjectBody {
  title?: string;
  slug?: string;
  description?: string;
  context?: string;
  problem?: string;
  solution?: string;
  techStack?: string;
  results?: string;
  category?: UpdateProjectBodyCategory;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  metric?: string | null;
  featured?: boolean;
  status?: string;
}